+++ addon/components/variable-number/number.ts
@@ -8,7 +8,6 @@ import {
   TextSelection,
 } from '@lblod/ember-rdfa-editor';
 import { action } from '@ember/object';
-import { tracked } from '@glimmer/tracking';
 import { inject as service } from '@ember/service';
 import intlService from 'ember-intl/services/intl';
 import { localCopy } from 'tracked-toolbox';
@@ -17,7 +16,8 @@ import {
   MINIMUM_VALUE_PNODE_KEY,
 } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/variable-plugin/utils/constants';
 import { isBlank } from '@ember/utils';
-import n2words from 'n2words';
+import { isNumber } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/strings';
+import { numberToWords } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/variable-plugin/utils/number-to-words';
 
 type Args = {
   getPos: () => number | undefined;
@@ -32,7 +32,6 @@ export default class VariableNumberPluginNumberComponent extends Component<Args>
   @localCopy('args.node.attrs.value', '') declare inputNumber: string;
   @localCopy('args.node.attrs.writtenNumber', false)
   declare writtenNumber: boolean;
-  @tracked errorMessage = '';
   @service declare intl: intlService;
   cursorPositionKeyDown: number | null = null;
 
@@ -46,13 +45,13 @@ export default class VariableNumberPluginNumberComponent extends Component<Args>
   get formattedNumber() {
     const value = this.node.attrs.value as string;
 
-    if (Number.isNaN(Number(value)) || value === null || value === '') {
+    if (!isNumber(value)) {
       return value;
     }
     if (!this.writtenNumber) {
       return value;
     } else {
-      return n2words(Number(value), { lang: 'nl' });
+      return numberToWords(Number(value), { lang: 'nl' });
     }
   }
 
@@ -78,41 +77,41 @@ export default class VariableNumberPluginNumberComponent extends Component<Args>
     this.args.updateAttribute('writtenNumber', !this.writtenNumber);
   }
 
-  validateAndSave() {
+  get errorMessage() {
     if (isBlank(this.inputNumber)) {
-      this.errorMessage = '';
-      this.args.updateAttribute('value', '');
-      return;
+      return '';
     }
-
     const number = Number(this.inputNumber);
     if (Number.isNaN(number)) {
-      this.errorMessage = this.intl.t('variable.number.error-not-number');
-      return;
+      return this.intl.t('variable.number.error-not-number');
     }
     const validMinimum = !this.minValue || number >= this.minValue;
     const validMaximum = !this.maxValue || number <= this.maxValue;
 
     if (!validMinimum || !validMaximum) {
       if (this.minValue && this.maxValue) {
-        this.errorMessage = this.intl.t(
-          'variable.number.error-number-between',
-          { minValue: this.minValue, maxValue: this.maxValue }
-        );
+        return this.intl.t('variable.number.error-number-between', {
+          minValue: this.minValue,
+          maxValue: this.maxValue,
+        });
       } else if (this.minValue) {
-        this.errorMessage = this.intl.t('variable.number.error-number-above', {
+        return this.intl.t('variable.number.error-number-above', {
           minValue: this.minValue,
         });
       } else if (this.maxValue) {
-        this.errorMessage = this.intl.t('variable.number.error-number-below', {
+        return this.intl.t('variable.number.error-number-below', {
           maxValue: this.maxValue,
         });
       }
-      return;
     }
+    return '';
+  }
 
-    this.errorMessage = '';
-    this.args.updateAttribute('value', this.inputNumber);
+  @action
+  validateAndSave() {
+    if (!this.errorMessage) {
+      this.args.updateAttribute('value', this.inputNumber);
+    }
   }
 
   @action
@@ -121,8 +120,8 @@ export default class VariableNumberPluginNumberComponent extends Component<Args>
     tr.setSelection(
       NodeSelection.create(
         this.args.controller.activeEditorState.doc,
-        this.args.getPos() as number
-      )
+        this.args.getPos() as number,
+      ),
     );
     this.args.controller.activeEditorView.dispatch(tr);
   }
@@ -155,7 +154,7 @@ export default class VariableNumberPluginNumberComponent extends Component<Args>
   setSelectionAt(pos: number) {
     const tr = this.args.controller.activeEditorState.tr;
     tr.setSelection(
-      TextSelection.create(this.args.controller.activeEditorState.doc, pos)
+      TextSelection.create(this.args.controller.activeEditorState.doc, pos),
     );
     this.args.controller.focus();
     this.args.controller.activeEditorView.dispatch(tr);
@@ -163,7 +162,7 @@ export default class VariableNumberPluginNumberComponent extends Component<Args>
 
   selectAfterNode() {
     this.setSelectionAt(
-      (this.args.getPos() as number) + this.args.node.nodeSize
+      (this.args.getPos() as number) + this.args.node.nodeSize,
     );
   }