+++ tests/dummy/app/controllers/regulatory-statement-sample.ts
@@ -12,6 +12,7 @@ import {
 } from '@lblod/ember-rdfa-editor/plugins/text-style';
 import {
   block_rdfa,
+  docWithConfig,
   hard_break,
   horizontal_rule,
   invisible_rdfa,
@@ -70,6 +71,10 @@ import {
   numberView,
 } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/variable-plugin/number';
 import { document_title } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/document-title-plugin/nodes';
+import {
+  templateCommentNodes,
+  templateCommentView,
+} from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/template-comments-plugin';
 export default class RegulatoryStatementSampleController extends Controller {
   @service declare importRdfaSnippet: ImportRdfaSnippet;
   @service declare intl: IntlService;
@@ -82,51 +87,50 @@ export default class RegulatoryStatementSampleController extends Controller {
     say: 'https://say.data.gift/ns/',
   };
 
-  get schema() {
-    return new Schema({
-      nodes: {
-        doc: {
-          content:
-            'table_of_contents? document_title? ((chapter|block)+|(title|block)+|(article|block)+)',
-        },
-        paragraph,
-        document_title,
-        repaired_block,
+  schema = new Schema({
+    nodes: {
+      doc: docWithConfig({
+        content:
+          'table_of_contents? document_title? ((chapter|block)+|(title|block)+|(article|block)+)',
+      }),
+      paragraph,
+      document_title,
+      repaired_block,
 
-        list_item,
-        ordered_list,
-        bullet_list,
-        placeholder,
-        ...tableNodes({ tableGroup: 'block', cellContent: 'block+' }),
-        date: date(this.config.date),
-        variable,
-        number: number,
-        ...STRUCTURE_NODES,
-        heading,
-        blockquote,
+      list_item,
+      ordered_list,
+      bullet_list,
+      ...templateCommentNodes,
+      placeholder,
+      ...tableNodes({ tableGroup: 'block', cellContent: 'block+' }),
+      date: date(this.config.date),
+      variable,
+      number: number,
+      ...STRUCTURE_NODES,
+      heading,
+      blockquote,
 
-        horizontal_rule,
-        code_block,
+      horizontal_rule,
+      code_block,
 
-        text,
+      text,
 
-        image,
+      image,
 
-        hard_break,
-        block_rdfa,
-        table_of_contents: table_of_contents(this.config.tableOfContents),
-        invisible_rdfa,
-        link: link(this.config.link),
-      },
-      marks: {
-        inline_rdfa,
-        em,
-        strong,
-        underline,
-        strikethrough,
-      },
-    });
-  }
+      hard_break,
+      block_rdfa,
+      table_of_contents: table_of_contents(this.config.tableOfContents),
+      invisible_rdfa,
+      link: link(this.config.link),
+    },
+    marks: {
+      inline_rdfa,
+      em,
+      strong,
+      underline,
+      strikethrough,
+    },
+  });
 
   get config() {
     return {
@@ -138,7 +142,7 @@ export default class RegulatoryStatementSampleController extends Controller {
           ],
           scrollContainer: () =>
             document.getElementsByClassName(
-              'say-container__main'
+              'say-container__main',
             )[0] as HTMLElement,
         },
       ],
@@ -176,23 +180,24 @@ export default class RegulatoryStatementSampleController extends Controller {
         interactive: true,
       },
       snippet: {
-        endpoint: 'http://localhost/sparql',
+        endpoint: 'https://dev.reglementairebijlagen.lblod.info',
       },
     };
   }
 
   @tracked rdfaEditor?: SayController;
   @tracked nodeViews: (
-    controller: SayController
+    controller: SayController,
   ) => Record<string, NodeViewConstructor> = (controller) => {
     return {
       variable: variableView(controller),
       table_of_contents: tableOfContentsView(this.config.tableOfContents)(
-        controller
+        controller,
       ),
       link: linkView(this.config.link)(controller),
       date: dateView(this.config.date)(controller),
       number: numberView(controller),
+      templateComment: templateCommentView(controller),
     };
   };
   @tracked plugins: Plugin[] = [
@@ -202,7 +207,7 @@ export default class RegulatoryStatementSampleController extends Controller {
       [space, hardBreak, paragraphInvisible, headingInvisible],
       {
         shouldShowInvisibles: false,
-      }
+      },
     ),
   ];
 
@@ -233,7 +238,7 @@ export default class RegulatoryStatementSampleController extends Controller {
     const presetContent =
       localStorage.getItem('EDITOR_CONTENT') ??
       `<div resource='http://localhost/test' typeof='say:DocumentContent'>Insert here</div>`;
-    controller.setHtmlContent(presetContent);
+    controller.initialize(presetContent);
     const editorDone = new CustomEvent('editor-done');
     window.dispatchEvent(editorDone);
   }